Refactor VoxelTraversal, extend WebGLStub for specs
